+++ bearc/src/compiler/hir/type_resolver.hpp
@@ -210,7 +210,7 @@ template <IsDefVisitor V> class TypeResolver {
         }
 
         auto maybe_size_exec = ComptExprSolver<V>{context, def_visitor}.solve_builtin_compt_expr(
-            fid, scope, type->type.arr.size_expr, builtin_type::usize, std::nullopt);
+            fid, scope, type->type.arr.size_expr, builtin_type::u64, std::nullopt);
 
         if (!maybe_size_exec.has_value()) {
             return OptId<TypeId>{};
